move Portal to the Default namespace
