+++ GeometryDashAPI/Levels/GameObjects/Specific/SpeedBlock.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using GeometryDashAPI.Attributes;
 using GeometryDashAPI.Levels.Enums;
+using GeometryDashAPI.Levels.GameObjects.Default;
 
 namespace GeometryDashAPI.Levels.GameObjects.Specific
 {
